Add contact note for Canonical employees
It's very common for SRU team members who are Canonical employees to get
private pings asking for SRU processing. But then, if we have questions,
those questions end up in private threads when they should be part of
the public conversation.

Otherwise we end up with information valuable to other developers being
withheld, and disenfranchise Ubuntu developers who are not Canonical
employees.

Initial pings on private channels should therefore generally not be
acceptable, but is difficult to deal with diplomatically when it
happens. Maybe this will help.