Temporarily drop Launchpad CI
Launchpad CI has never passed, and its presence and failure is confusing
to contributors. We should fix CI, but in the meantime, as long as we're
ignoring the failures completely, we might as well avoid confusion by
not running CI that is known to fail in the first place.

See: LP: #2081852