explanation: SRU refusal is an acceptable outcome
We recently had a case where a contributor seemed to believe that it was
the SRU team's reponsibility to shape their proposal into something
acceptable. There are two flaws with this mindset, both of which are
addressed here:

1) It's the SRU Driver's role to provide an acceptable SRU and the
appropriate expertise, not the SRU Team's role to turn a proposal into
something acceptable or to become a subject matter expert.

2) If nobody can provide appropriate risk analysis and mitigation,
refusal of the SRU is an acceptable outcome.