index.rst: add process overview and diagram
The landing page for SRU documentation should contain a basic
explanation of the process. Developers should not need to dig through
the docs for a tl;dr on the SRU process.

The explanation/pipeline.rst document has this information. Include it
in the main index.rst, and add a diagram to illustrate the pipeline.