Add Documentation of Special Cases from Wiki
This is the section about "Documentation of Special Cases" from the
wiki, converted to reStructuredText in the same way as the exceptions
itself.

It's original history up to the state of this content is retained
in the out of tree StableReleaseUpdates.moin. Originally up to v389
by racb in commit 2d15834e, the remaining versions are
continued, merged and present in 3690d6d8 of this effort.