Exceptions: deduplicate unused labels
The `make run` did not warn about duplicates when a label
was not used in any reference. Direct sphinx build did.

Being unused, we could as well remove the labels that got
created out of headers in the conversion, but we can apply
the same "add prefix" approach we did for the used labels
and keep them around to allow easy cross-references.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>