phased-updates: first pass of technical updates
That documentation is written like a live document, more or less playing
the role of a bug tracker with open questions and future work. This
doesn't fit a documentation page, so I've removed most of those entries,
and corrected what I could so that at least the technical information in
that page is correct. This still needs an overhaul to better fit modern
documentation standards.