DMB: Clarify and extend applications
Based on the most common questions that we got from applicants
we extended on the list of expectations from Robie. Framing it
more in the way an applicant would read and consume it.
Extending by examples and details on the context.

Furthermore this split allows to add the often requested
list of knowledge skills and how it maps to different
application levels.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>