package-specific: update linux-firmware release process
If linux-firmware is released to both updates and security at the same time, it
causes a lot of strain on the archive servers, because:
- it's a large package (over 500MiB in size)
- the security pocket is not phased
- all ubuntu installations point at the archive server for security updates,
  without using a mirror

The kernel team was contacted, and confirmed that linux-firmware SRUs can
first be released to updates, and later copied to security. In fact, that's
how the kernel SRUs themselves are released: first updates, later security.

See matrix (short) thread:

https://matrix.to/#/!slcYrFqWytcsbPzMxA:ubuntu.com/$UlulzAKm8ZkamZzH16Mu8jSCy8px6G3RxDqhzcquMcE?via=ubuntu.com&via=matrix.org&via=matrix.debian.social