PGP/Signing/Uploads - connect these intertwinded topics
Currently the setup of GPG keys, the definition of a signature and
a signing key as well as their usage for an upload is independent.
This is due to coming from different sources, but consuming that as
someone new to the project is quite hard.

Allow readers to find one from the other by extending their
description and linking between them.