PGP key storage recommendations
For a long time everyone has done a lot on their own to keep PGP
keys safe, and to be fair many already do what I now suggest to
officially recommend.
But to step up the integrity of the project we need to start somewhere
and here that is by outlining clear recommendations that can
become deeply tested, documented and eventually mandatory.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>