PGP: list known missing aspects visible to the reader
Discussions can happen in chat or mailing lists or elsewhere.
Proper documentation issue tracking for improvement can be done
in issues, but in this case there is a high chance for some aspects
the reader expects and starts to search in vain. Hence track the
top level entries of that directly on the page for reader
awareness.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>