MIR: Identify who is responsible
We used to have just the flowchart and table and tried to balance
this between readable and easy, as well as detailed and complete.
But it was neither as it was always a tradeoff.

Now that we have established a simpler explanation in the overview
we can tackle issue #130 and go all in to make the flowchart
fully complete (all states have bug assignee info, all transitions
are indexed, ownership is colorized). Furthermore as we did for the
states this establishes a table for the transitions to allow more
text than what fits well in the flowchart.

Now the chart has a few crossing lines, but everying is much
more correct and clearly states who we expect to handle any
given state and transition.

Fixes: #130

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>