workaround for md format: rendering links with arguments
Sadly the current build chain is broken, report #313 was only one
instance of a general failure.
- Links in rst are not affected
- Links in markdown files are affected
  - links inside of escaping blocks like ```foo``` are not affected
  - links inside of template blocks like ```notes ...``` are affected

Ideally the build would be fixed and this change reverted.
If that is impossible the link checker needs to learn the above
conditions and fail on any link with arguments that is in markdown
files while not being in an escaping block.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>