rendering: mitigate MyST Parser issue 1028
This is post-processing HTML files to mitigate
https://github.com/executablebooks/MyST-Parser/issues/1028

Without that fix URLs fail to work in the browsers merging arguments and
values together and silently break the usage of those links.
Contributors do not and should not need to know about this.
It does to by unescaping &amp;amp; to &amp; which is the correct
format.

Affected are links in markdown like:
  [text](https://wiki.ubuntu.com/Home?action=foo&option=bar)
Without the fix they become:
  href="https://wiki.ubuntu.com/Home?action=foo&amp;amp;option=bar"
with the fix they are correct like:
  href="https://wiki.ubuntu.com/Home?action=foo%amp;option=bar"

To be minimally invasive this is:
  - only running for html builders
  - only changing text in href attributes
  - only affecting parametrized URLs that follow to an ?
  - only affecting &amp;amp; (over just &amp;) which is a
    more clear indication of that bug
  - only writing on changed content

This change can be reverted once we rebased to a fixed myst-parser.

Fixes: #332

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>